parser: Separate nodes into multiple files
